requirements.txt
gTTS==2.3.1
newspaper3k==0.2.8
nltk==3.8.1
